🚜Rename: img_create 폴더 생성
image generate에 관여하는 파일을 img_create폴더 내부로 옮김
